Bump to Bazel 4.0 (#237)
* Bump to Bazel 4.0

This PR bumps rules_ios to Bazel 4.0 and includes misc changes. Many
have been addressed in other PRs. It also makes some changes to the
bundling rules, which were required for rules_apple toolchain feature.

Note:
it also rebases the rules_swift fork to the latest revision and pushed
tag for it. We hope to eventually upstream the changes for IWBV2, which
would be a followup.