Enable test action associations in schemes
A convention many projects utilize is to have test
targets run as a test action for a scheme where the build
target is not itself a test target, including the
default project generated in Xcode.

MyFramework <-- framework
MyFrameworkTests <-- bundle.unittests

This hopes to enable adding test targets and environment
variables to schemes created for any build target.