Wrap ios_application dependencies in framework
Currently when using this rule set with an application organized around
`apple_framework_packaging` the rules_ios transition is missing from the
app sources. Source files picked up in the `ios_application`a globs are
sent through `apple_library`. This defaults to wrapping the code
apple_framework_packaging to apply the transition.

Alternatives considered: it may be ideal to just remove the "library"
feature from this macro to simplify the implementation and eliminate the
need to partition out kwargs. That may be ideal longer term but will
require changes in other repos.