minimum_deployment_os_version / `-target` stopgap
It adds the ability to set `minimum_deployment_os_version` and pulls in
the latest sha of `rules_apple`.

It stopgaps target specific evaluation xcconfig issues. When I tried to
build an app with target specifc xcconfigs with settings like
`SWIFT_PLATFORM_TARGET_PREFIX`, I end up getting invalid `-target` flags
in compiler invocations which were incomplete. `rules_ios`'s xcconfig
evaluator is not a full implementation of xcconfig and conflicts with
Bazel's settings here. It doesn't currenlty evaluate on a per archicture
basis but aligning the 2 may be a longer term exercise.

This PR is related to #225 but doesn't fix it. Longer term, we'll need a
way to either remove `platforms` or find a way to pin override them in a
way like #258 takes care of.