Scaffolding for Xcode configuration repo rule
Sets up the scaffolding to address https://github.com/bazel-ios/rules_ios/issues/269

Provide a fast way to configure Xcode and set versions. Bazel currently
contains an Xcode auto configure mechanism that doesn't work so well
with `rules_ios` and resulted in performance issues when used with
Xcode's GUI and the CLI.

Bazel requires a user to set the Xcode verson flag. Today an Xcode
project generator or wrapper must interact with the `xcode_vesrion` flag
and make make assumptions about how it works: e.g.  XCHammer sets
`xcode_version` based on `xcodebuild -version` for speed, and
`rules_ios` currently triggers re-analysis because it runs Bazel query (
https://github.com/bazel-ios/rules_ios/issues/269 ). This pattern is not
great for maintainablility, or performance.

All of this is replaced by an auto-configuation program on SPM / `swift
run`. The program will manage both Bazel's Xcode version flag and the
overall. It uses `swift run` / SPM for incremental speed and simplicity
as we can't Bazel build a that's an input to a repo rule. For now, the
basic logic of what it will do is written in a comment and all of this
is scaffolding, so it doesn't do anything yet.