Virtual hermetic frameworks POC
Add the ability to virtualize the framework structure. swift and clang
compile as frameworks but the files don't move and we load them via VFS.

This is a working POC - for perf benchmarks only and as a reference of
the work.

Hacked this up to see how it impacts build performance in a big app.
Cut clean build time in half in a big hybrid app from rules_ios
baseline. More interestingly, hot top level compiler invocations are
down nearly .5 orders of magnitude.