Virtual hermetic frameworks (#277)
* Virtual hermetic frameworks

Add the ability to virtualize the framework structure. swift and clang
compile as frameworks but the files don't move and we load them via VFS.
The VFS is llvm's in-memory file system declared by JSON. VFS is already
in use but this uses it in a very different way.

This fixes many issues:
1. we don't hit O(n) deps includes - removing thousands of framework
search paths
2. we don't have to "clean" frameworks. This feature added quite a bit
of overhead and isn't necesary since it's declarative
3. imports are fully again. traditionally a build system uses -F on a
build directory. This causes problems with reproducible builds.

It cut clean build time in half in a big hybrid app from rules_ios
baseline. More interestingly, hot top level compiler invocations are
down nearly .5 orders of magnitude.


This required adding a way to build `ios_application` `srcs as a framework.
If this feature is on, the apps need to have it set since they don't
build as a framework

Remaining tasks:
- address xcodeproj integration
- duplicate test suite to run with this on/off
- address xcframework imports ( right now we're using -F )


Relates to: #211