Fix conflict with umbrella headers (#278)
The line to determine umbrella headers will cause issues in some
scenarios. This PR reverts to the orginal, albeit more nuanced logic.
Additionally correct the comment which would lead us to believe that
this is OK to disregard.

Will work on getting this explified in rules_ios as a followup.