Use absolute vfs paths via LLVM `VirtualFileSystem`
There are issues when using non absolute-path VFS outside of the
compilation directory, even when setting the `working-directory`. We
can't use absolute paths in this file in Bazel for many reasons. To work
around this, use VFS paths relative to the VFS root via compiler
internals.

Internal to swift and clang - LLVM `VirtualFileSystem` object can
serialize paths relative to the absolute path of the overlay. This
requires the paths are relative to the overlay. While deriving the
in-memory tree roots, it pre-pends the prefix of the `vfsoverlay` path
to each of the entries.

In practice, this is required to use rules_ios vfsoverlay with Xcode's
indexer and SourceKit: e.g. in Bazel Xcode project generators like the
one in rules_ios or XCHammer.