Reduce VFS assumptions about swiftmodules
The previous code relied on assumptions about swiftmodules which may be
invalid, depending on the input.

Now:
- It puts the name of the parent in the framework if it's a
  swiftmodule-directory. This handlt es user provided directory
  swiftmodules. The code now checks if its a `.swiftmodule` because
  previously, it could take _any_ swiftmodule and give it the name of the
  parent: e.g. the build directory.

- It takes the name of the swiftmodule and puts it into the root as
  opposed to computing it.

It can fall back to the code path with assumptions to improve useability
of the rule. There would need to be further refactoring remove this:
Ideally remove this "apple libary only" codepath in the future.