Fixes default swiftmodule include issue (#297)
By default `rules_swift` propagates swiftmodules to includes when they
exist as a dep. For now, we need to propagate these as they were.

Additionally, it factors out SwiftInfo building to another function.
This is orthogonal to the change, but something we came up with in the
continued effort to make the framework large framework method a bit
easier to follow than originally.