[virtual-frameworks] Fix non-modular framework umbrella
In order to correctly idenify some imports as a framework import, clang
needs the entire subtree under an imported framework. The previous path
triggered an edge case in some vendor frameworks through certian import
paths.

Builds out the VFS subtrees for a given set of paths in this situation,
and then adds them to the VFS.

This manifests in one very odd permutations of obscure vendor frameworks
when using umbrella headers the headers. I suspect it's more
generalizeable than the previous logic