Multiple architecture support in apple_framework_packaging
This fixes #186.

When building for multiple archs, we need to uses ctx.split_attr.[attr] to access the attributes which attach the apple_common.multi_arch_split transition. This allows us to use the artifacts built for the correct arch in the current target. To identify the correct configuration of a target at build time, we use ctx.fragments.apple.single_arch_cpu and tx.fragments.apple.single_arch_platform.platform_type to find the current architecture and Apple platform type.

Please take a look at transitions_and_multi_arch_support.md for the explanation of Bazel transtions and how to use it to support multi arches.