Conditionally compile with transitive VFS
Conditionally adds VFS all targets, which right now has conditional side
effect to compile them all as virtual frameworks. In the interim we want
to include the VFS in order to pickup deps that are virtual frameworks.

It also includes a stopgap hack for issues with merging imported
frameworks. The inline comment may be addressed

Additionally, it includes a github action to build the entire test suite
with virtual frameworks to pickup regressions.