Fix external VFS paths
Currently, the VFS rule formulates a path relative to the BUILD file
path, to replicate the "exec path" in the java rules. Now this path
formulation include the repository name for external repos. Of course,
it isn't returned by `build_file_path`. We didn't have a test case and
thanks to @jschear, it's now up and running

Fixes #307