Supply FrameworkInfo from ad-hoc framework imports
Today, there are effectivly 2 different ways that frameworks are being
imported. The first one is `apple_library` where it creates a framework
import for adjacent frameworks. The other case is calling
`apple_*_framework_import` from the rules. For now, create framework
import for the latter, consider alignment of these codepaths longer
term.

This commit just instantiates a `FrameworkInfo` here so it's passed
upwards.