[RFC] A rule to update ARM binaries to run on M1 arm64 sim
This PR is an RFC on an _additional_ rule to rules_ios that adds the ability to
update the Mach-o header on imported libraries and frameworks to get arm64
binaires running on Apple silicon simulator. For the demo, it's added in
`app.bzl` but will need a feature flag and more.

Why bother doing this? Well some apps have many dependencies which could take
along time on vendors or other parties to update. Because the M1 chip has the
same ISA as ARM64, most binaries will run transparently. Most iOS developers
code is high level enough and isn't specifc to a device or simulator. There are
many caveats and eceptions but getting it running is better than nothing. ( e.g.
`TARGET_OS_SIMULATOR` )

This solves the problem at the build system level with the power of bazel. The
idea is pretty straight forward:
1. collect all imported paths
2. update the macho headers with Apples vtool and arm-to-sim
3. update the linker invocation to use the new libs

Now it updates all of the inputs automatically - the action can be taught to do
all of this conditionally if necessary.

Additionally the apsect was generating the action but I moved it into the rule.
Mainly because you can't make an executable input to an aspect to scheudle
actions there.  This has an interesting propery: you get a single path for
framework lookups at linktim.  Today with bazel and rules_ios ld has an
O(MFrameworks*NFrameworkPaths) search requirmenet. This is an additional
consequence and worth comparing link time impovements but is a future exercise.

There are some minor followups, but seeing what people think about it