Fork Bazel's Xcode autoconfigure for usability improvements and debugging
Fork and improve Bazel's xcode configuration code. This is not a measure to fix
the problem, but to debug it and unblock CI. Github actions is incrediblly
flaky but Bazel should deal with a flaky host better and this is a weak point
(https://github.com/bazel-ios/rules_ios/issues/361)

Bazel puts this error in the build file which is a problem on the CI because it
swallows the error and it requires a user to clean --expunge. Bazel requires
the user to expunge because it thinks the repository was successfully written.
Probably a better pattern is to fail, but that won't shed light on the problem.

This is not just broken for CI but local developers. The plan is to use
rules_ios CI to reproduce this and gather more information then propose a fix
when we know more info.