Quote ARCH variable in update_static.sh
Shellcheck is a static analysis tool that highlights syntactic issues
that lead to common shell bugs. It highlighted that `ARCH` is not quoted
in this script, and when I quote it, the script does not work correctly.

Opening this as a PR at @jmarino's request so that he can see the
change.