Improve M1 Simulator Bazel 5 CPU handling
This PR adds support for the Bazel 5 CPU type, `ios_sim_arm64` during
transition to platforms and disables the Bazel 4 x86_64 -> arm64 hack.
This allows us to use the new CPU type with the ARM64 middleman to run
on the M1 Simulator with device deps. Finally, we don't need to have an
extra step to run on M1 Simulator like on Bazel 4.

Additionally:
- bump Bazel off RC and rules_apple for requiste support
- disable Bazel 4 toolchain given Bazel 5 fixes