LLDB iOS high level testing rules and tests
Given the complexity of LLDB, it's not enough to just test LLDB in
rules_ios and call it done. Ideally should be tested under each app.
Finally, LLDB often breaks across Xcode updates without any
intervention, so this makes it a bit easier to catch and give us a head
start on fixing it or even flagging it to related parties.

This PR implements `lldb_test` so you can run LLDB inside of a bazel
repo and verify debugger functionality end to end. `lldb_test` is a rule
that you can specify the tests.

It includes a python library for constructing iOS LLDB command line
automation, oriented around testing. It orchestrates a simulator thread,
LLDB thread, to run simulators and debuggers concurrently. It transmits
information between Bazel and LLDB with JSON files. The Bazel rule just
plugs the python program into Bazel, but this could work for other build
systems.

Note: there are a couple of FIXMEs but this atleast gets it running on
CI