Generate lldbinits from installer for LLDB test
Feeds the LLDB init test from the rules into the xcodeproj

It sxport Xcode build settings JSON to an env file, sources it and then
runs `lldb-settings.sh` Gist is to dump showBuildSettings to replicate
xcode runscript actions

This is a quick hack for testing, and the env file is not reproducible
because the absolute paths.  Theoretically the LLDBInit should be
relative, the LLDBInit is less necessary it's current form with "virtual
frameworks".

Additionally silence output that is large on a big app