Fixes import_middleman when ios_multi_cpus=x86_64,arm64
The select method uses the configuration of the current target's configuration. The outgoing edge transition doesn't affect the select method. Therefore, it's possible that an import_middleman target with a non-sim_arm64 configuration is analyzed during the analysis phase, which causes crash.

An alternative solution is to change transition_support.apple_rule_transition to output //command_line_option:ios_multi_cpus with a single cpu. This is a bigger change and might affect downstream usage.