[Xcode 13.3] patch `WorkspacePath` onto test failures
In Xcode 13.3 XCTFail stoped sending absolute source file paths to
`-[XCTSourceCodeLocation initWithFileURL:lineNumber:]` so reading the
`DO_NOT_BUILD_HERE` from the parent of URL stopped working.

Better, prepend the `info.plist` for the corresponding bundle to the
relative path if it doesn't already exist. On 13.* it works the same.