Fix `platforms` / os version analysis phase caching
The nature of this attriute led to it being cached indeffinitly in the
analysis phase. In other words, previously, once Bazel would set it and
then it was cached until the analysis cache was invalidated either
directly or indirectly. Forward the value to `minimum_os_version`
directly, which is is part of the analysis cache key for C++ rules