Add support for project options in `legacy_xcodeproj` (#553)
# Summary
We override the `Text Settings` in our non-`legacy_xcodeproj` XcodeGen config:
```YML
options:
  # Text indent
  usesTabs: false
  indentWidth: 2
  tabWidth: 2
```
Add support for `options` to be overridden via the `project_options_overrides` in `legacy_xcodeproj`. It might seem weird that this requires an entire new `rule` impl but this is done since XcodeGen won’t process `”2”` correctly from a `string_dict` for arguments like `indentWidth` or `tabWidth`. `_ProjectOptionsInfo `, a new provider, allows the values to be forwarded to the JSON payload as the properly encoded `int` and `bool` values:
```JSON
{
    "options": {
        "createIntermediateGroups": true,
        "defaultConfig": "Debug",
        "groupSortPosition": "none",
        "settingPresets": "none",
        "usesTabs": false,
        "indentWidth": 2,
        "tabWidth": 2
    },
}
```