[Xcode 13.3.1] Flip on in Swift PCH fix
Flips PCH feature to mitigate broken PCH ingestion in some cases as a
default. There are issues w/o this even for rules_ios test cases on
13.3.1 and probably apps had to turn that on for this bump. Because
there are no known down-sides and it was somewhat involved to debug
and create the PCH feature - turn it on a sensible default.