Remove INFOPLIST_FILE setting from the legacy xcodeproj rule
In Xcode 14, we see that the invocation of `builtin-infoPlistUtility`
sometimes will come after the 'Build with bazel' script phase. When this
happens, the plist value will get overwritten by the one we've specified
in our rules. As we always provide a base Info.plist (as required by
rules_apple) if none is specified, this stub seems unnecessary.