Fix infoplist exclusivitty / config inhertiance
The original version required that infoplists_by_build_setting and
xcconfig_by_build_setting are mutually exclusive which was problematic
for people who had different condtions. To remidate this problem, let
build configuration pick _one_: move xcconfig_by_build_setting into a
select() statement as-is so it is selected by configuration

Finally remove the "select" statement inheritance. From an API
perspectie a select isn't inherited so we will only inherit from
`xcconfig`. This is a breaking change if you relied on the previous
behavior ( which was only a couple days ago )