Feed custom `module_map` to compile actions
At some point in recent updates/bumps we hit a a regression which caused
custom module maps to no longer be propagated to compile actions. This
broke remote execution and caching.

Ensures the `input` maps also are sent to `objc_library` - prior to
overriding it in the proceeding line of code.