[LTS][Bazel 5] Provide minimum LTS support
From time to time we'll evaluate supporting this to allow us to all run
on HEAD and keep pushing / working together with minimal dammage.

Detect if the user is on Bazel 5.0 and fallback to our tag of
`rules_apple` and corresponding API. Because we depend on some of the
more stable internals, this isn't such a big deal for now. Add CI
vetting and a README note on the status.

Outside of this, a minor fix make it work for me: disable strict version
validation we added around deps.

Bazel 6.0 has a number of open questions that will be answered later