[M1][xcodeproj] CLI to update xcodeproj rules tests
This PR adds a way for people to run tests which test rules_ios
xcodeproj rules tests.  Anytime anyone looks at a change that impacts
transitions ( including myself ) it causing the world more confusion to
figure out the magic encantation of the test program and rosetta. You
can easily run the tests with `--update` if you just care to fix the
transitions.

I added assertions for rosetta if you work on arm64 - now it just fails
to prevent people getting stuck on this.

Update the fixtures with:
```
arch -arch x86_64 /bin/bash -l -c ./tests/xcodeproj-tests.sh --update
```

Fixes #598

Tested and working end to end on the latest cut of macOS Montery and
Xcode 14.1.

It needs a login shell or `xcodebuild` doesn't work.