Remove extra filegroups (#711)
**Note:** This is a port of https://github.com/bazel-ios/rules_ios/pull/548 after the changes in https://github.com/bazel-ios/rules_ios/pull/708.

This creates a number of extra rules that can go away - it's effectivly unused

This change is done for simplify - not heap size. This said, it cuts an entire category from the target graph: Because this removes 3 for each rule,  we'll eliminate this many pointers from memory:
```
O(Napple_framework * Kheaders * 3)
```

After this change - you'll end up with O(Nresource groups) filegroups. For me, in larger build - the measured impact is a reduction of ~10k filegroups -> 89