Drop support for Bazel 5 & rules_apple 2 (#834)
This stops testing Bazel 5/rules_apple 2 in CI and bumps the support
Bazel version in MODULE.bazel. It does not include any of the Bazel 5
related cleanup that is now possible. This is for an upcoming version
4.0.0.

Adds a table like the following to the read me:

## Supported Bazel versions

| Bazel release | Minimum supported rules version | Final supported
rules version

|:-------------------:|:-------------------------:|:-------------------------:
| 6.* | 2.0.0 | current
| 5.* | 1.0.0 | 3.2.2
| 4.* | 1.0.0 | 1.0.0

## Supported [rules_apple](https://github.com/bazelbuild/rules_apple)
versions

| rules_apple release | Minimum supported rules version | Final
supported rules version

|:-------------------:|:-------------------------:|:-------------------------:
| 3.* | 3.* | current
| 2.* | 2.* | 3.2.2
| 1.* | 1.0.0 | 3.2.2