Move resource generation below framework vfs
This moves resource generation (and being added to `deps`) below the `framework_vfs` creation which uses the `deps`

With the change in #879 to add the vfs to the `swiftc_inputs`  instead of the `deps` of the `swift_library` the `framework_vfs` was now also forwarding the resources (because it gets added to `deps` after generation) to other rules (like `rules_xcodeproj`) which collect extra files from this the `swiftc_inputs` attr.

AFAIK resources aren't needed for the `framework_vfs` rules and so this change should resolve the new `rules_xcodeproj` incompatibility while making the `deps` to the `framework_vfs` more scoped.