Allow library tool resource handling to be disabled (#928)
This allows the wrap resource in filegroups and resource bundle
generators to be disabled completely. It allows users to just pass in
`data` and have the rule just forward that to the underlying libraries.

This cleans up the code a bit too since we use `data` directly instead
of `module_data` which had to be converted into a list in a few places.

I added a test showcasing where this ability would be required. It fails
without this change and passes when disabling the library tools