Add objc_library binary to outputs of apple_framework_packaging rule when virtualized frameworks feature is enabled (#931)
This fixes an issue where building an `apple_framework_packaging` rule
with the VFS feature enabled wouldn't trigger objc compilation for the
framework.

When the VFS feature is disabled, the binaries from the `objc_library`
and `swift_library` targets are merged using libtool, and that binary is
included in the outputs of the `apple_framework_packaging` rule ([see
packaging
action](https://github.com/bazel-ios/rules_ios/blob/f7dcd0c4f90985495bba517dc8cc7d86ce7f7632/rules/framework/framework_packaging.py#L19)).

When the VFS feature is enabled, all of the packaging actions are
skipped, and the binaries from the `objc_library` and `swift_library`
targets aren't merged. One of the input binaries is arbitrarily included
in the rule's output files -- for a target containing swift and objc
sources, this would be the `swift_library` binary.

The PR adds both input binaries to the rule outputs when the VFS feature
is enabled, so building an `apple_framework_packaging` rule triggers
both objc and swift compilation actions.